Lab1-b solved
